Backend/requirements.txt
asgiref==3.5.2
attrs==23.1.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
bidict==0.23.1
bleach==6.1.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==38.0.4 
fastjsonschema==2.19.0
Flask==2.3.3
flask-abort==0.0.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
flask-marshmallow==1.2.0
flask-pyjwt==1.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.3
marshmallow==3.21.0
marshmallow-sqlalchemy==1.0.0
mistune==3.0.2
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2 
packaging==23.2
pandocfilters==1.5.0
platformdirs==4.1.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-engineio==4.9.0
python-socketio==5.11.2
pyzmq==25.1.2
referencing==0.32.0
requests==2.31.0
rpds-py==0.13.2
simple-websocket==1.0.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.25
sqlparse==0.4.3
sympy==1.12
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.0
typing_extensions==4.9.0
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.1
wsproto==1.2.0
